# docs/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Bastion",
    "description": "ERP para pyme española. Contrato de la versión 1 de la API, bajo /api/v1/{modulo}/{recurso}. Se genera al compilar desde los controladores y los DTO: no se escribe a mano, y el cliente del frontal se genera de él.",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/organizacion/almacenes/{id}": {
      "get": {
        "tags": [
          "Almacenes"
        ],
        "summary": "Devuelve un almacén.",
        "operationId": "Almacenes_Obtener",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador del almacén.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AlmacenDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlmacenDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlmacenDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          { }
        ]
      },
      "put": {
        "tags": [
          "Almacenes"
        ],
        "summary": "Cambia los datos de un almacén.",
        "operationId": "Almacenes_Modificar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador del almacén.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModificarAlmacenDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ModificarAlmacenDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ModificarAlmacenDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AlmacenDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlmacenDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlmacenDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "412": {
            "description": "Precondition Failed",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "428": {
            "description": "Precondition Required",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          { }
        ]
      },
      "delete": {
        "tags": [
          "Almacenes"
        ],
        "summary": "Bloquea un almacén. No lo borra.",
        "description": "Cada movimiento de existencias apunta a su almacén para siempre: borrar la fila rompería el\nhistórico de valoración, que no se puede reconstruir después.",
        "operationId": "Almacenes_Bloquear",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador del almacén.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "412": {
            "description": "Precondition Failed",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "428": {
            "description": "Precondition Required",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          { }
        ]
      }
    },
    "/api/v1/organizacion/almacenes/{id}/desbloqueo": {
      "post": {
        "tags": [
          "Almacenes"
        ],
        "summary": "Devuelve un almacén bloqueado a la operativa.",
        "description": "Sin puerta HTTP en el 0.4 por lo mismo que el desbloqueo de empresa, y abierta hoy por lo\nmismo: detrás de su permiso, que no es el de bloquear.",
        "operationId": "Almacenes_Desbloquear",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador del almacén.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          { }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AlmacenDto": {
        "required": [
          "id",
          "empresaId",
          "codigo",
          "nombre",
          "direccion",
          "tipo"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identificador del almacén.",
            "format": "uuid"
          },
          "empresaId": {
            "type": "string",
            "description": "Empresa a la que pertenece (R8).",
            "format": "uuid"
          },
          "codigo": {
            "type": "string",
            "description": "Código del almacén, en mayúsculas."
          },
          "nombre": {
            "type": "string",
            "description": "Nombre con el que se le conoce."
          },
          "direccion": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "description": "Dónde está, o nulo si es virtual o de tránsito.",
                "$ref": "#/components/schemas/DireccionDto"
              }
            ]
          },
          "tipo": {
            "type": "string",
            "description": "Tipo de almacén, como texto."
          }
        },
        "description": "Un almacén, tal como sale de la API."
      },
...
      "DireccionDto": {
        "required": [
          "calle",
          "codigoPostal",
          "poblacion",
          "pais"
        ],
        "type": "object",
        "properties": {
          "calle": {
            "maxLength": 70,
            "minLength": 0,
            "type": "string",
            "description": "Nombre de la vía."
          },
          "numero": {
            "maxLength": 16,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ],
            "description": "Número, portal, piso. Opcional: hay direcciones que no lo tienen."
          },
          "codigoPostal": {
            "maxLength": 16,
            "minLength": 0,
            "type": "string",
            "description": "Código postal, tal como lo escribe el país de la dirección."
          },
          "poblacion": {
            "maxLength": 35,
            "minLength": 0,
            "type": "string",
            "description": "Población."
          },
          "subdivision": {
            "maxLength": 35,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ],
            "description": "Provincia, región o estado. Opcional: no todos los países la usan."
          },
          "pais": {
            "maxLength": 2,
            "minLength": 2,
            "type": "string",
            "description": "País en ISO 3166-1 alfa-2, dos letras."
          }
        },
        "description": "Una dirección estructurada, en los seis campos que exige R17."
      },
...
      "ModificarAlmacenDto": {
        "required": [
          "nombre",
          "tipo"
        ],
        "type": "object",
        "properties": {
          "nombre": {
            "maxLength": 120,
            "minLength": 0,
            "type": "string",
            "description": "Nombre con el que se le conoce."
          },
          "direccion": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "description": "Dónde está, o nulo si es virtual o de tránsito.",
                "$ref": "#/components/schemas/DireccionDto"
              }
            ]
          },
          "tipo": {
            "type": "string",
            "description": "Tipo de almacén, como texto."
          }
        },
        "description": "Lo que se puede cambiar de un almacén."
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "testigoDeAcceso": {
        "type": "http",
        "description": "Testigo de acceso de quince minutos, en la cabecera Authorization. Lo emite POST /api/v1/identidad/sesiones y lo renueva POST .../sesiones/renovacion, que no lo lee de aquí sino de una cookie HttpOnly. El testigo de refresco no aparece en este contrato a propósito: el navegador no tiene que poder leerlo.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}